+++ CHANGES.md
@@ -1,3 +1,7 @@
+master
+------
+* Http#with_response (through Chainable)
+
 0.2.0
 -----
 * Request and response objects